requirements.txt
aiogram==3.18.0
alembic==1.14.0
cachetools==5.5.0
python-dotenv==1.0.1
SQLAlchemy==2.0.36
asyncpg==0.30.0